fix: preserve empty-spef object identity in RCX→STA handoff; clear branch-new lint; split sizer tests
Correctness (AC-4): EngineFlow.create_step_workspaces forwarded the
predecessor's
spef with `pre_step.output.spef or []`, which substitutes a fresh list
for the common
empty case. The legacy `get("spef", [])` forwarded the predecessor's own
list object.
Restore the exact contract with a direct typed handoff `=
pre_step.output.spef`, and
parameterize test_rcx_to_sta_spef_transfer over empty + nonempty,
asserting both content
and object identity.

Static gate (AC-5): remove the two branch-only F401 `WorkspaceStep`
imports left after
the variant migration (ecc_dreamplace/checklist.py,
ecc_sizer/test_module.py). A
line-anchored base(main)/HEAD ruff differential across the 56 touched
files now shows
zero branch-new diagnostics (the 6 apparent hits on
data/workspace/__init__.py are
byte-identical pre-existing lines in main's relocated workspace.py).
Pre-existing
whole-repo debt left untouched per the relative gate.

Review bar: split test/tools/ecc_sizer/test_module.py (713 lines, over
the 700-line
guideline) — move the runner/flow tests to
test/tools/ecc_sizer/test_runner.py and share
_workspace/_subflow_states/_sizer_runtime via a local _sizer_helpers.py.
Both files now
under 700 lines (395 / 283); moved tests collect and pass.

Verification: full pytest 1035 passed; four-tool build smoke PASS;
layout.py ruff+pyright
clean; git diff --check clean; no new cast/type-ignore.
